Change button setting for beexybox
The current settings do not work fine with the beexybox. The buttons
do not correspond nicely with the stimuli. Now the left stimulus
corresponds with the left button of a 3-button beexybox and the
right stimulus corresponds with the right button of the beexybox.
The lower button is the "continue" button.